withNamedThemes: `theme` prop can be a 2-item array
I’ve added support for a `theme` prop like `[‘baseTheme’, themeObjectOrName]`, for ordered usage of a base theme and an override theme. This behavior is intended to respect the `composeTheme` prop.

There’s also a third argument to the `@withNamedThemes()` decorator: a theme object or, more pertinently, the name of the default theme. Previously I was using a string-type defaultProps.theme within a decorated component, but this led to that string being merged with the final theme object passed to the component.